refactor(cli): five pure-move splits out of the largest files
Production code only moves; every test stays where it was and reaches
the moved items through the parent's imports. Verified with the
line-multiset check from #674 (`perf-tools/puremove.py`).

  config/mod.rs        -> config/dotenv.rs (requote, load_dotenv_filtered)
                          config/perms.rs  (create_config_dir, check_permissions*,
                                            set_dir_permissions*)
  dashboard/state.rs   -> state/sync.rs    (sync_from_run_state, now_secs,
                                            looks_stale; a child module so it
                                            keeps the struct's private fields)
  serve/runs.rs        -> runs/matching.rs (apply_search and the highlight
                                            functions)
  runstate.rs          -> runstate/dashboard_log.rs, runstate/force.rs
  daemon/script_host.rs-> script_host/permissions.rs, script_host/http_limits.rs

One doc comment was sitting on the wrong function: `load_dotenv_filtered`'s
paragraph had been written above `requote`'s, so rustdoc attached both to
`requote`. It now sits on the function it describes.
