+++ crates/leviath-cli/src/daemon/script_host.rs
@@ -26,6 +26,16 @@ use tokio::process::Command as TokioCommand;
 use crate::config::{ScriptPermission, ScriptToolPermissions, ToolPolicy};
 use crate::daemon::sandbox_manager::SandboxManager;
 
+mod http_limits;
+mod permissions;
+use http_limits::*;
+#[cfg(test)]
+pub(crate) use http_limits::{REDIRECT_MIRROR, lock_redirect_mirror};
+pub use http_limits::{
+    set_local_network_allowed, set_script_http_max_per_host, set_script_http_timeout,
+};
+pub use permissions::{effective_script_permissions, resolve_script_permissions};
+
 /// The resolved allow/deny decision for each of the five side-effecting host
 /// functions, computed once at spawn from the config's `[tool_script_permissions]`
 /// and the agent's own tool permissions (for the `inherit` cases).
@@ -43,122 +53,6 @@ pub struct ScriptAllow {
     pub write_file: bool,
     /// Whether `env_var` may run.
     pub env_var: bool,
-}
-
-/// Resolve `[tool_script_permissions]` into concrete allow/deny booleans.
-///
-/// `Allow`/`Deny` map directly. `Inherit` means:
-/// - `read_file` / `write_file` / `shell`: permitted only when the agent's resolved policy for
-///   the equivalent built-in (`resolve_builtin`) is [`ToolPolicy::Allow`]. This
-///   is evaluated once against the entry stage's permission layers; a later
-///   stage's `tool_permissions` do not re-gate a script's host calls.
-/// - `http_get` / `http_post` / `env_var`: permitted (no built-in equivalent to
-///   inherit from, and the tool itself is still gated by Layers 1/2/4).
-///
-/// `resolve_builtin` is a `&dyn Fn` (not `impl Fn`) so this function has a single
-/// monomorphization; otherwise each distinct caller closure type gets its own
-/// copy of the `net`/`filelike` match arms, and coverage is attributed
-/// per-instantiation (each only exercises the arms that caller hits).
-pub fn resolve_script_permissions(
-    perms: &ScriptToolPermissions,
-    resolve_builtin: &dyn Fn(&str) -> ToolPolicy,
-) -> ScriptAllow {
-    let net = |p: ScriptPermission| match p {
-        ScriptPermission::Allow | ScriptPermission::Inherit => true,
-        ScriptPermission::Deny => false,
-    };
-    let filelike = |p: ScriptPermission, builtin: &str| match p {
-        ScriptPermission::Allow => true,
-        ScriptPermission::Deny => false,
-        ScriptPermission::Inherit => resolve_builtin(builtin) == ToolPolicy::Allow,
-    };
-    ScriptAllow {
-        http_get: net(perms.http_get),
-        http_post: net(perms.http_post),
-        env_var: net(perms.env_var),
-        read_file: filelike(perms.read_file, "read_file"),
-        write_file: filelike(perms.write_file, "write_file"),
-        shell: filelike(perms.shell, "shell"),
-    }
-}
-
-/// Map a `[tool_script_permissions]` string to a [`ScriptPermission`]. An
-/// unrecognized value yields `None` (the field is left at the global default) -
-/// parsed by hand (not via `Deserialize`) so every arm is deterministically
-/// covered, without pulling in serde's unexercised visitor machinery.
-fn parse_script_permission_str(s: &str) -> Option<ScriptPermission> {
-    match s {
-        "allow" => Some(ScriptPermission::Allow),
-        "deny" => Some(ScriptPermission::Deny),
-        "inherit" => Some(ScriptPermission::Inherit),
-        _ => None,
-    }
-}
-
-/// How restrictive a script permission is, for clamping.
-///
-/// `Allow` (unconditional) is the loosest; `Inherit` still requires the agent's
-/// own policy for the equivalent built-in to permit the call; `Deny` is the
-/// tightest.
-fn script_restrictiveness(p: ScriptPermission) -> u8 {
-    match p {
-        ScriptPermission::Allow => 0,
-        ScriptPermission::Inherit => 1,
-        ScriptPermission::Deny => 2,
-    }
-}
-
-/// The effective `[tool_script_permissions]` for an agent: the user's global
-/// config with the agent's own blueprint `[tool_script_permissions]` overlaid
-/// per field - but **only where the manifest is more restrictive**.
-///
-/// Agents ship their own `.rhai` tool scripts, so it is reasonable for a
-/// manifest to say "this agent never needs `shell`". It is not reasonable for it
-/// to say the opposite: a manifest that could set `shell = "allow"` over a user's
-/// global `deny` meant installing an agent was enough to overrule the machine's
-/// configuration. So a manifest may tighten a field and never loosen it, the same
-/// rule [`crate::tools::resolve_policy`] applies to `[tool_permissions]`.
-///
-/// Parsed CLI-side (these types live in the CLI config, not `leviath-core`),
-/// mirroring `parse_blueprint_mcp_servers`.
-pub fn effective_script_permissions(
-    global: &ScriptToolPermissions,
-    manifest_toml: &str,
-) -> ScriptToolPermissions {
-    let mut eff = global.clone();
-    // `toml::from_str`, not `manifest_toml.parse::<toml::Value>()`. In toml 1.x
-    // `FromStr for Value` parses a single *value*, not a document - so a real
-    // manifest starting with `[agent]` reads as an array literal followed by
-    // junk and fails. It still compiles, so the change is silent; the tests are
-    // what caught it.
-    let Ok(value) = toml::from_str::<toml::Value>(manifest_toml) else {
-        return eff;
-    };
-    let Some(table) = value
-        .get("tool_script_permissions")
-        .and_then(|v| v.as_table())
-    else {
-        return eff;
-    };
-    // For each key the agent set to a recognized value, keep whichever of the
-    // two is stricter.
-    let apply = |key: &str, slot: &mut ScriptPermission| {
-        if let Some(p) = table
-            .get(key)
-            .and_then(|v| v.as_str())
-            .and_then(parse_script_permission_str)
-            && script_restrictiveness(p) > script_restrictiveness(*slot)
-        {
-            *slot = p;
-        }
-    };
-    apply("http_get", &mut eff.http_get);
-    apply("http_post", &mut eff.http_post);
-    apply("shell", &mut eff.shell);
-    apply("read_file", &mut eff.read_file);
-    apply("write_file", &mut eff.write_file);
-    apply("env_var", &mut eff.env_var);
-    eff
 }
 
 /// The raw I/O a [`DaemonScriptHost`] performs, behind a seam so the host's
@@ -501,164 +395,6 @@ static HTTP_CLIENT: std::sync::LazyLock<reqwest::blocking::Client> =
 /// author as the same opaque sentence.
 fn error_chain(e: &(dyn std::error::Error + 'static)) -> String {
     leviath_providers::rhai_provider::host::error_chain(e)
-}
-
-/// Concurrent script-tool HTTP requests allowed per host; `0` is unbounded.
-///
-/// A process-wide mirror of `[limits] script_http_max_per_host`, for the same
-/// reason [`ALLOW_LOCAL_REDIRECTS`] is one: [`HTTP_CLIENT`] is process-wide and
-/// the request path has no handle on the config.
-static HTTP_MAX_PER_HOST: std::sync::atomic::AtomicUsize = std::sync::atomic::AtomicUsize::new(4);
-
-/// Apply `[limits] script_http_max_per_host` for this process.
-pub fn set_script_http_max_per_host(max: usize) {
-    HTTP_MAX_PER_HOST.store(max, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// In-flight script-tool requests per host, and the condvar waiters park on.
-static IN_FLIGHT: std::sync::LazyLock<std::sync::Mutex<std::collections::HashMap<String, usize>>> =
-    std::sync::LazyLock::new(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-static IN_FLIGHT_FREED: std::sync::Condvar = std::sync::Condvar::new();
-
-/// A permit to have one request in flight against `host`, released on drop.
-///
-/// A condvar rather than a `tokio::sync::Semaphore` because script tools run on
-/// `spawn_blocking` threads with no runtime handle to await on.
-struct HostPermit(Option<String>);
-
-impl HostPermit {
-    /// Take a permit for `url`'s host, blocking while that host is at its cap.
-    ///
-    /// A URL with no host (and an unbounded cap) takes no permit at all, so the
-    /// unlimited setting costs nothing.
-    fn acquire(url: &str) -> Self {
-        let max = HTTP_MAX_PER_HOST.load(std::sync::atomic::Ordering::Relaxed);
-        let Some(host) = host_of(url).filter(|_| max > 0) else {
-            return Self(None);
-        };
-        let mut counts = IN_FLIGHT
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        while counts.get(&host).copied().unwrap_or(0) >= max {
-            counts = IN_FLIGHT_FREED
-                .wait(counts)
-                .unwrap_or_else(std::sync::PoisonError::into_inner);
-        }
-        *counts.entry(host.clone()).or_insert(0) += 1;
-        Self(Some(host))
-    }
-}
-
-impl Drop for HostPermit {
-    fn drop(&mut self) {
-        let Some(host) = self.0.take() else {
-            return;
-        };
-        let mut counts = IN_FLIGHT
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        // `or_insert(1)` rather than `if let Some(..)`: acquire always inserted
-        // the key and this is its only remover, so the missing case cannot
-        // happen - and written as a branch it was a region no test could enter.
-        let remaining = counts.entry(host.clone()).or_insert(1);
-        *remaining = remaining.saturating_sub(1);
-        if *remaining == 0 {
-            counts.remove(&host);
-        }
-        drop(counts);
-        IN_FLIGHT_FREED.notify_all();
-    }
-}
-
-/// The host part of `url`, lowercased. `None` when it does not parse.
-fn host_of(url: &str) -> Option<String> {
-    reqwest::Url::parse(url)
-        .ok()?
-        .host_str()
-        .map(str::to_lowercase)
-}
-
-/// Seconds a script tool's HTTP request may take; `0` leaves the client's own
-/// deadline in charge.
-///
-/// Applied per request rather than on [`HTTP_CLIENT`] because that static is
-/// built lazily at first use, which may be before or after the config is read.
-/// A per-request timeout also wins over the client's, so this is the value that
-/// actually governs.
-static HTTP_TIMEOUT_SECS: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(30);
-
-/// Apply `[limits] script_http_timeout_secs` for this process.
-pub fn set_script_http_timeout(secs: u64) {
-    HTTP_TIMEOUT_SECS.store(secs, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// The configured per-request deadline, if one is set.
-fn script_http_timeout() -> Option<Duration> {
-    match HTTP_TIMEOUT_SECS.load(std::sync::atomic::Ordering::Relaxed) {
-        0 => None,
-        secs => Some(Duration::from_secs(secs)),
-    }
-}
-
-/// How many extra attempts a script tool's request gets after a transport
-/// failure that looks transient.
-const SCRIPT_HTTP_RETRIES: u32 = 2;
-
-/// Whether *redirect hops* may land on loopback / private / link-local
-/// addresses.
-///
-/// The authoritative check is [`DaemonScriptHost::allow_local_network`], a plain
-/// field on the host. This atomic exists only because [`HTTP_CLIENT`] is
-/// process-wide and its redirect callback runs inside reqwest with no access to
-/// the host that started the request. `[security] allow_local_network` is a
-/// machine-wide switch, so one value per process is the right granularity -
-/// but keep the field authoritative and this a mirror of it, not the reverse:
-/// global mutable state read by the main check would make every test that
-/// touches it race with every test that doesn't.
-///
-/// Defaults to `false`, so a path that forgets to initialize it is the safe one.
-static ALLOW_LOCAL_REDIRECTS: std::sync::atomic::AtomicBool =
-    std::sync::atomic::AtomicBool::new(false);
-
-/// The lock every test that touches [`ALLOW_LOCAL_REDIRECTS`] must hold.
-///
-/// The atomic is process-wide, so in a test binary - where everything runs in
-/// one process, in parallel - a test that writes it races every test that reads
-/// it, and the loser sees the other test's value with no hint that is what
-/// happened. A refusal test quietly succeeds, or a permitted-hop test is refused
-/// and blames the thing it was actually checking.
-///
-/// It lives next to the atomic rather than inside one module's test block
-/// because the writers are not all in one module: `setup_daemon_host_with`
-/// mirrors the config into it at daemon start-up, so a test that stands up a
-/// host is a writer too - and that was the one with no idea it had to take this.
-/// An async mutex rather than a `std` one because the writers await: standing up
-/// a daemon host is an `async fn`, so a `std` guard held across it is
-/// `clippy::await_holding_lock` - and the lint is right, the guard would be
-/// pinned to whatever thread the future resumed on. This one also cannot be
-/// poisoned, which matters for a lock every test in three modules takes: a test
-/// that panicked holding it has already said why, and failing every later test
-/// as well would bury that one real failure.
-#[cfg(test)]
-pub(crate) static REDIRECT_MIRROR: tokio::sync::Mutex<()> = tokio::sync::Mutex::const_new(());
-
-/// Take the [`REDIRECT_MIRROR`] lock from synchronous code.
-///
-/// Async callers should `REDIRECT_MIRROR.lock().await` instead; this is for the
-/// plain `#[test]`s and for a `spawn_blocking` body, neither of which can.
-#[cfg(test)]
-pub(crate) fn lock_redirect_mirror() -> tokio::sync::MutexGuard<'static, ()> {
-    REDIRECT_MIRROR.blocking_lock()
-}
-
-/// Apply `[security] allow_local_network` to redirect following for this process.
-pub fn set_local_network_allowed(allow: bool) {
-    ALLOW_LOCAL_REDIRECTS.store(allow, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// The current value of the [`ALLOW_LOCAL_REDIRECTS`] switch.
-fn local_network_allowed() -> bool {
-    ALLOW_LOCAL_REDIRECTS.load(std::sync::atomic::Ordering::Relaxed)
 }
 
 impl RealScriptIo {
