+++ crates/leviath-cli/src/runstate.rs
@@ -29,6 +29,13 @@
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
+mod dashboard_log;
+mod force;
+#[cfg(test)]
+use dashboard_log::*;
+pub use dashboard_log::{append_dashboard_log, append_dashboard_log_to, dashboard_log_path};
+pub use force::{ForceCancelOutcome, force_cancel, force_cancel_in, force_error_in};
+
 // The plain run-state data types (RunMeta, RunStatus, the snapshot structs, and
 // the per-stage records) live in `leviath_core::run_meta`. Re-exported here so
 // `crate::runstate::RunMeta` / `runstate::RunMeta` call sites across the cli
@@ -332,92 +339,6 @@ pub fn run_dir(run_id: &str) -> PathBuf {
     runs_dir().join(run_id)
 }
 
-/// Inner implementation of `dashboard_log_path`, parameterised so it can be
-/// tested without touching the process-global env. All callers go through
-/// `dashboard_log_path`.
-fn dashboard_log_path_from(env_override: Option<&str>) -> PathBuf {
-    if let Some(path) = env_override {
-        return PathBuf::from(path);
-    }
-    leviath_core::paths::data_dir()
-        .unwrap_or_default()
-        .join("dashboard.log")
-}
-
-/// Path to the persistent dashboard activity log (~/.leviath/dashboard.log).
-///
-/// Honours the `LEVIATH_DASHBOARD_LOG_PATH` override when set (tests use it via
-/// `isolate_runs_dir_for_test`); otherwise resolves the real home-relative
-/// path. This function only *computes* a `PathBuf` - it never writes - so both
-/// arms are safe to exercise directly in tests. The write side
-/// ([`append_dashboard_log`] and `Dashboard::add_log`) is what must stay off
-/// the user's real log in tests: `append_dashboard_log`'s own tests set the
-/// override, and `Dashboard` carries an injected log path (a temp dir under
-/// `make_test_dashboard`) so no dashboard-input test ever appends to the real
-/// `~/.leviath/dashboard.log`.
-pub fn dashboard_log_path() -> PathBuf {
-    match std::env::var("LEVIATH_DASHBOARD_LOG_PATH") {
-        Ok(path) => dashboard_log_path_from(Some(&path)),
-        Err(_) => dashboard_log_path_from(None),
-    }
-}
-
-/// Append a timestamped line to the persistent dashboard activity log at the
-/// default [`dashboard_log_path`]. Silently ignores I/O errors - best-effort.
-pub fn append_dashboard_log(msg: &str) {
-    append_dashboard_log_to(&dashboard_log_path(), msg);
-}
-
-/// Append a timestamped line to the dashboard activity log at an explicit
-/// `path`. Silently ignores I/O errors - the dashboard log is best-effort.
-///
-/// The path is a parameter so `Dashboard` can inject a test-isolated log
-/// location, guaranteeing no dashboard-input test appends to the user's real
-/// `~/.leviath/dashboard.log` (see [`dashboard_log_path`]).
-pub fn append_dashboard_log_to(path: &Path, msg: &str) {
-    append_dashboard_log_capped(path, msg, DASHBOARD_LOG_MAX_BYTES);
-}
-
-/// The dashboard log is capped at this size; once the live file reaches it, the
-/// file is rolled (see [`roll_log_if_over_cap`]) so it can't grow without bound
-/// across a long-lived daemon's lifetime.
-const DASHBOARD_LOG_MAX_BYTES: u64 = 5 * 1024 * 1024;
-
-/// Append with an explicit cap (the public entry points use
-/// [`DASHBOARD_LOG_MAX_BYTES`]; tests pass a small cap to exercise rolling).
-fn append_dashboard_log_capped(path: &Path, msg: &str, max_bytes: u64) {
-    use std::io::Write;
-    // Ensure the parent directory exists (first-run case).
-    if let Some(parent) = path.parent() {
-        let _ = std::fs::create_dir_all(parent);
-    }
-    roll_log_if_over_cap(path, max_bytes);
-    if let Ok(mut file) = leviath_sys::open_private_append(path) {
-        let timestamp = chrono::Local::now().format("%Y-%m-%d %H:%M:%S");
-        let _ = writeln!(file, "{} {}", timestamp, msg);
-    }
-}
-
-/// The path the rolled (previous-generation) log is moved to: `<name>.1`.
-fn rolled_log_path(path: &Path) -> PathBuf {
-    let mut name = path.as_os_str().to_owned();
-    name.push(".1");
-    PathBuf::from(name)
-}
-
-/// Roll the live log to `<name>.1` once it reaches `max_bytes`, replacing any
-/// existing rolled file, so the live file restarts empty and at most one
-/// previous generation is retained (bounded ~2×cap on disk). Best-effort - a
-/// failed rename just leaves the log to keep growing rather than erroring.
-fn roll_log_if_over_cap(path: &Path, max_bytes: u64) {
-    let over = std::fs::metadata(path)
-        .map(|m| m.len() >= max_bytes)
-        .unwrap_or(false);
-    if over {
-        let _ = std::fs::rename(path, rolled_log_path(path));
-    }
-}
-
 /// How many random bits go in a run ID's suffix, rendered as 12 hex digits.
 /// Collisions only matter within one wall-clock second for one agent name, so 48
 /// bits is many orders of magnitude more than needed while staying short enough
@@ -595,124 +516,6 @@ pub fn looks_abandoned(
     }
     let moved_at = meta.last_progress_at.unwrap_or(meta.updated_at);
     now.saturating_sub(moved_at) > STALE_AFTER_SECS
-}
-
-/// The outcome of forcing a run to a terminal state on disk.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ForceCancelOutcome {
-    /// The run was live on disk and is now recorded terminal.
-    Terminated,
-    /// The run was already finished; nothing was written.
-    AlreadyTerminal,
-    /// No run directory with that id exists.
-    NoSuchRun,
-    /// The directory exists but its metadata could not be rewritten.
-    WriteFailed,
-}
-
-impl ForceCancelOutcome {
-    /// Whether the id named a run at all - i.e. whether the cancel had a target,
-    /// regardless of whether it needed to write anything.
-    pub fn found_run(&self) -> bool {
-        !matches!(self, Self::NoSuchRun)
-    }
-}
-
-/// Force a run's on-disk metadata to `Cancelled`, in the runs dir resolved from
-/// the environment. See [`force_cancel_in`].
-pub fn force_cancel(run_id: &str) -> ForceCancelOutcome {
-    force_cancel_in(&run_dir(run_id), leviath_core::duration::now_secs())
-}
-
-/// Force the run in `run_dir` to `Cancelled`, stamping `updated_at` with `now`.
-///
-/// This is the floor under every kill path: it needs nothing but the filesystem,
-/// so it works for a run the daemon can't rebuild (blueprint deleted, metadata
-/// corrupt, died mid-spawn) and for a run whose daemon is gone entirely. Both
-/// the daemon's force-terminator seam and `lev cancel --force` route here so
-/// there is one definition of "terminated on disk".
-///
-/// A directory whose `meta.json` is missing or unparseable still gets a minimal
-/// `Cancelled` record written: such a run is otherwise skipped by `list_runs`,
-/// which makes it invisible *and* permanent.
-pub fn force_cancel_in(run_dir: &Path, now: i64) -> ForceCancelOutcome {
-    force_terminal_in(run_dir, RunStatus::Cancelled, None, now)
-}
-
-/// Force the run in `run_dir` to `Error` with `message`, stamping `updated_at`.
-///
-/// For the spawn that never became a run. The spawner stakes out the run
-/// directory and writes a `Starting` placeholder *before* building the agent, so
-/// a spawn that fails leaves something to diagnose - but `Starting` is not
-/// terminal, so that placeholder went on claiming the run was alive for ever,
-/// showing up in `lev ps` and the dashboard with nothing behind it (issue #190).
-/// Recording the failure where the placeholder is turns it into an answer.
-pub fn force_error_in(run_dir: &Path, message: &str, now: i64) -> ForceCancelOutcome {
-    force_terminal_in(run_dir, RunStatus::Error, Some(message.to_string()), now)
-}
-
-/// Rewrite the run in `run_dir` to a terminal `status`, attaching `error` when
-/// there is something to say. Shared by [`force_cancel_in`] and
-/// [`force_error_in`] so "terminated on disk" has one implementation.
-fn force_terminal_in(
-    run_dir: &Path,
-    status: RunStatus,
-    error: Option<String>,
-    now: i64,
-) -> ForceCancelOutcome {
-    if !run_dir.is_dir() {
-        return ForceCancelOutcome::NoSuchRun;
-    }
-    let run_id = run_dir
-        .file_name()
-        .map(|n| n.to_string_lossy().into_owned())
-        .unwrap_or_default();
-    let terminated = match read_meta_from(run_dir) {
-        Ok(meta) if is_terminal_status(&meta.status) => return ForceCancelOutcome::AlreadyTerminal,
-        Ok(meta) => RunMeta {
-            status,
-            updated_at: now,
-            // Keep whatever the run had already recorded when there is nothing
-            // new to say (the cancel path).
-            error: error.clone().or(meta.error),
-            ..meta
-        },
-        // Unreadable metadata: synthesize just enough to record the outcome. The
-        // run id is the directory name, which is the one field always recoverable.
-        Err(_) => RunMeta {
-            status,
-            updated_at: now,
-            error: Some(
-                error
-                    .clone()
-                    .unwrap_or_else(|| "run metadata was unreadable; cancelled".to_string()),
-            ),
-            ..RunMeta::new(
-                run_id.clone(),
-                run_id,
-                String::new(),
-                String::new(),
-                None,
-                String::new(),
-                0,
-            )
-        },
-    };
-    match write_meta_to(run_dir, &terminated) {
-        Ok(()) => ForceCancelOutcome::Terminated,
-        Err(e) => {
-            // Formatted outside the macro: a method call inside a `%field` is
-            // only evaluated when a subscriber visits the value, so it would go
-            // unexercised under the tests' no-op subscriber.
-            let path = run_dir.display().to_string();
-            tracing::warn!(
-                run_dir = %path,
-                error = %e,
-                "could not force a run to a terminal state on disk"
-            );
-            ForceCancelOutcome::WriteFailed
-        }
-    }
 }
 
 /// Read run metadata out of an explicit run directory (the daemon works from its
